Rename the Types files.
