--- src/components/Formik/FormikSelectInput/FormikSelectInput.jsx
@@ -1,43 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import SelectInput from '../../SelectInput/SelectInput';
-
-const propTypes = {
-  field: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-    onBlur: PropTypes.func.isRequired,
-    onChange: PropTypes.func.isRequired,
-    value: PropTypes.oneOfType([
-      PropTypes.object,
-      PropTypes.string,
-      PropTypes.arrayOf(PropTypes.object),
-    ]),
-  }).isRequired,
-  form: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
-};
-
-const FormikSelectInput = (
-  {
-    field: {
-      name,
-      onBlur,
-      onChange,
-      value,
-    },
-    form: { touched, errors },
-    ...props
-  },
-) => (
-  <SelectInput
-    name={name}
-    onBlur={onBlur}
-    onChange={onChange}
-    value={value}
-    error={touched[name] && errors[name]}
-    {...props}
-  />
-);
-
-FormikSelectInput.propTypes = propTypes;
-
-export default FormikSelectInput;
