+++ src/components/SelectInput/SelectInputTypes.tsx
@@ -1,0 +1,13 @@
+import { ValueType, OptionTypeBase } from 'react-select';
+
+export interface SimulatedEventPayload {
+  target: {
+    name: string;
+    value: ValueType<OptionTypeBase>;
+  };
+}
+
+export interface Option {
+  value: string;
+  label: string;
+}
